Moved the ParseServer class into a new file
