+++ addParseCloud.js
@@ -1,0 +1,41 @@
+var Parse = require('parse/node').Parse;
+var httpRequest = require('./httpRequest');
+
+function getClassName(parseClass) {
+  if (parseClass && parseClass.className) {
+    return parseClass.className;
+  }
+  return parseClass;
+}
+
+
+module.exports = function addParseCloud() {
+  Parse.Cloud.Functions = {};
+  Parse.Cloud.Triggers = {
+    beforeSave: {},
+    beforeDelete: {},
+    afterSave: {},
+    afterDelete: {}
+  };
+  Parse.Cloud.define = function(functionName, handler) {
+    Parse.Cloud.Functions[functionName] = handler;
+  };
+  Parse.Cloud.beforeSave = function(parseClass, handler) {
+    var className = getClassName(parseClass);
+    Parse.Cloud.Triggers.beforeSave[className] = handler;
+  };
+  Parse.Cloud.beforeDelete = function(parseClass, handler) {
+    var className = getClassName(parseClass);
+    Parse.Cloud.Triggers.beforeDelete[className] = handler;
+  };
+  Parse.Cloud.afterSave = function(parseClass, handler) {
+    var className = getClassName(parseClass);
+    Parse.Cloud.Triggers.afterSave[className] = handler;
+  };
+  Parse.Cloud.afterDelete = function(parseClass, handler) {
+    var className = getClassName(parseClass);
+    Parse.Cloud.Triggers.afterDelete[className] = handler;
+  };
+  Parse.Cloud.httpRequest = httpRequest;
+  global.Parse = Parse;
+}
